Rename the file in common that is called evaulation - to spell correctly. Fixes #53
