+++ cal_ratio_trainer/training/evaluate_training.py
@@ -11,7 +11,7 @@ import pandas as pd
 from keras import Model
 from keras.src.utils import np_utils
 from matplotlib import pyplot as plt
-from cal_ratio_trainer.common.evaulation import (
+from cal_ratio_trainer.common.evaluation import (
     normalize_to_one,
     plot_roc_curve,
     signal_llp_efficiencies,
